452: gamma_function.GetGamma

+++ a_mini/projects/MiniAgg/03_SpanGenerators/IGammaFunction.cs
@@ -23,7 +23,7 @@ namespace PixelFarm.Agg
 {
     public interface IGammaFunction
     {
-        double GetGamma(double x);
+        float GetGamma(float x);
     }
