+++ a_mini/projects/MiniAgg/04_Scanline/0_ScanlineSpan.cs
@@ -6,19 +6,19 @@ namespace PixelFarm.Agg
     {
         public short x;
         public short len;
-        public int cover_index;
+        public short cover_index;
 
         public ScanlineSpan(int x, int cover_index)
         {
-            this.x =(short) x;
+            this.x = (short)x;
             this.len = 1;
-            this.cover_index = cover_index;
+            this.cover_index = (short)cover_index;
         }
         public ScanlineSpan(int x, int len, int cover_index)
         {
             this.x = (short)x;
             this.len = (short)len;
-            this.cover_index = cover_index;
+            this.cover_index = (short)cover_index;
         }
 #if DEBUG
         public override string ToString()
